Various improvements around the spell memorization interface.
Instead of applying changes directly to the player and then rebuilding the memorized spell list, changes are made to the list and the whole list is applied to the player instead. This allows for gaps in the memorized spell list, while the actual stored list on the player is densly packed.