Separate creation of the ConditionSpec objects and their configuration
This is required to fix issues with conditions that reference themselves or each other in their handlers. (i.e. Bane and Bless cancel each other). Previously these references would be null since the field was still running its initializer when the handlers were configured. Now the objects will all be created first, and then the configuration will be applied afterwards in the ConditionRegistry.